+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
 
 **Changed**
 
+- #1352 Make timeit to not display args by default
 - #1330 Make guards to not rely on review history
 
 **Removed**
@@ -20,6 +21,7 @@ Changelog
 
 **Fixed**
 
+- #1344 Handle inline images in Results Interpretation
 - #1336 Fix result capture date inconsistency
 - #1334 Number of analyses are not updated after modifying analyses in a Sample
 - #1319 Make api.get_review_history to always return a list